UserId saving with the notes feature added

+++ src/services/note.service.js
@@ -1,8 +1,8 @@
 import Note from '../models/notes.model';
 import bcrypt from 'bcrypt';
 
-export const getAllNotes = async () => {
-  const data = await Note.find();
+export const getAllNotes = async (UserID) => {
+  const data = await Note.find(UserID);
   return data;
 };
 
@@ -11,25 +11,20 @@ export const createNote = async (body) => {
       return data;
   };
 
-export const getNote = async (id) => {
-    const data = await Note.findById(id);
+export const getNote = async (id, UserID) => {
+    const data = await Note.findById({_id: id, UserID: UserID});
     return data;
   };
 
-export const updateNote = async (_id, body) => {
-    const data = await Note.findByIdAndUpdate(
-      {
-        _id
-      },
-      body,
-      {
-        new: true
-      }
-    );
+export const updateNote = async (id, body) => {
+    const data = await Note.findOneAndUpdate({ _id: id, UserID: body.UserID }, 
+      body,{
+      new: true
+    });
     return data;
   };
 
-export const deleteUser = async (id) => {
-    await Note.findByIdAndDelete(id);
+export const deleteNote = async (id, UserID) => {
+    await Note.findByIdAndDelete({_id: id, UserID: UserID});
     return '';
   };
